+++ internal/client/client.go
@@ -85,16 +85,6 @@ func (clnt *Client) SendToServer() error {
 
 		body, _ := json.Marshal(metric)
 
-		/*
-			var compressedBody bytes.Buffer
-			zw := gzip.NewWriter(&compressedBody)
-			defer zw.Close()
-
-			if _, err := zw.Write(body); err != nil {
-				return fmt.Errorf("compressing %s: %w", metric.ID, err)
-			}
-		*/
-
 		req, err := http.NewRequest(http.MethodPost, url, bytes.NewReader(body))
 		if err != nil {
 			return fmt.Errorf("creating request for %s: %w", metric.ID, err)
