+++ internal/server/middleware/compress.go
@@ -1,0 +1,71 @@
+package middleware
+
+import (
+	"bytes"
+	"compress/gzip"
+	"io"
+	"net/http"
+	"strings"
+)
+
+func DecompressMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.Header.Get("Content-Encoding") != "gzip" {
+			next.ServeHTTP(w, r)
+			return
+		}
+
+		bodyBytes, err := io.ReadAll(r.Body)
+		if err != nil {
+			http.Error(w, "bad body", http.StatusBadRequest)
+			return
+		}
+
+		gr, err := gzip.NewReader(bytes.NewReader(bodyBytes))
+		if err != nil {
+			r.Body = io.NopCloser(bytes.NewReader(bodyBytes))
+			next.ServeHTTP(w, r)
+			return
+		}
+		defer gr.Close()
+
+		r.Body = gr
+
+		next.ServeHTTP(w, r)
+	})
+}
+
+func CompressMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if !strings.Contains(r.Header.Get("Accept-Encoding"), "gzip") {
+			next.ServeHTTP(w, r)
+			return
+		}
+
+		grw := &gzipResponseWriter{ResponseWriter: w}
+		defer grw.Close()
+
+		next.ServeHTTP(grw, r)
+	})
+}
+
+type gzipResponseWriter struct {
+	http.ResponseWriter
+	writer *gzip.Writer
+}
+
+func (w *gzipResponseWriter) Write(b []byte) (int, error) {
+	if w.writer == nil {
+		w.Header().Set("Content-Encoding", "gzip")
+		gz := gzip.NewWriter(w.ResponseWriter)
+		w.writer = gz
+	}
+	return w.writer.Write(b)
+}
+
+func (w *gzipResponseWriter) Close() error {
+	if w.writer != nil {
+		return w.writer.Close()
+	}
+	return nil
+}
