--- temp_file
@@ -1,157 +1,0 @@
-{
-  "Alloc": {
-    "id": "Alloc",
-    "type": "gauge",
-    "value": 488888
-  },
-  "BuckHashSys": {
-    "id": "BuckHashSys",
-    "type": "gauge",
-    "value": 6937
-  },
-  "Frees": {
-    "id": "Frees",
-    "type": "gauge",
-    "value": 166
-  },
-  "GCCPUFraction": {
-    "id": "GCCPUFraction",
-    "type": "gauge",
-    "value": 0
-  },
-  "GCSys": {
-    "id": "GCSys",
-    "type": "gauge",
-    "value": 1708720
-  },
-  "GetSet11": {
-    "id": "GetSet11",
-    "type": "counter",
-    "delta": 2133651157
-  },
-  "GetSet13": {
-    "id": "GetSet13",
-    "type": "gauge",
-    "value": 194799.82652026275
-  },
-  "HeapAlloc": {
-    "id": "HeapAlloc",
-    "type": "gauge",
-    "value": 488888
-  },
-  "HeapIdle": {
-    "id": "HeapIdle",
-    "type": "gauge",
-    "value": 2801664
-  },
-  "HeapInuse": {
-    "id": "HeapInuse",
-    "type": "gauge",
-    "value": 1032192
-  },
-  "HeapObjects": {
-    "id": "HeapObjects",
-    "type": "gauge",
-    "value": 1182
-  },
-  "HeapReleased": {
-    "id": "HeapReleased",
-    "type": "gauge",
-    "value": 2768896
-  },
-  "HeapSys": {
-    "id": "HeapSys",
-    "type": "gauge",
-    "value": 3833856
-  },
-  "LastGC": {
-    "id": "LastGC",
-    "type": "gauge",
-    "value": 0
-  },
-  "Lookups": {
-    "id": "Lookups",
-    "type": "gauge",
-    "value": 0
-  },
-  "MCacheInuse": {
-    "id": "MCacheInuse",
-    "type": "gauge",
-    "value": 14400
-  },
-  "MCacheSys": {
-    "id": "MCacheSys",
-    "type": "gauge",
-    "value": 15600
-  },
-  "MSpanInuse": {
-    "id": "MSpanInuse",
-    "type": "gauge",
-    "value": 24480
-  },
-  "MSpanSys": {
-    "id": "MSpanSys",
-    "type": "gauge",
-    "value": 32640
-  },
-  "Mallocs": {
-    "id": "Mallocs",
-    "type": "gauge",
-    "value": 1348
-  },
-  "NextGC": {
-    "id": "NextGC",
-    "type": "gauge",
-    "value": 4194304
-  },
-  "NumForcedGC": {
-    "id": "NumForcedGC",
-    "type": "gauge",
-    "value": 0
-  },
-  "NumGC": {
-    "id": "NumGC",
-    "type": "gauge",
-    "value": 0
-  },
-  "OtherSys": {
-    "id": "OtherSys",
-    "type": "gauge",
-    "value": 684503
-  },
-  "PauseTotalNs": {
-    "id": "PauseTotalNs",
-    "type": "gauge",
-    "value": 0
-  },
-  "PollCount": {
-    "id": "PollCount",
-    "type": "counter",
-    "delta": 55
-  },
-  "RandomValue": {
-    "id": "RandomValue",
-    "type": "gauge",
-    "value": 0.32567062932470936
-  },
-  "StackInuse": {
-    "id": "StackInuse",
-    "type": "gauge",
-    "value": 360448
-  },
-  "StackSys": {
-    "id": "StackSys",
-    "type": "gauge",
-    "value": 360448
-  },
-  "Sys": {
-    "id": "Sys",
-    "type": "gauge",
-    "value": 6642704
-  },
-  "TotalAlloc": {
-    "id": "TotalAlloc",
-    "type": "gauge",
-    "value": 488888
-  }
-}
